Seed the database with the provided values
Built the scripts to seed that database for initial calculations.  This
meant changing the migration because I had made the deleted_at a NOT NULL
column.  This along with fixing the docker-compose allows us to dump in
initial data that we can use to test the offer decision logic once built.