Add LoanApplicationController and stub application submit API
Created the LoanApplicationController with the submit route.  For parity
added the equivalent route into the UserController.

This allows for the API to look up the loan application based on the
application or the user.  We also protect against applications that
do not exist.

Next is to build the decision logic