Simplified conversion from numeric to number
Sadly the implementation of TypeORM with Postgres returns numeric data
types as string values.  The supported way to handle this is to add in
a transformer that converts the value.  The first attempt used the library
decimal.js.  That was not needed and could be stripped down to just a
parseFloat.

I could also correct the casing of the name of the transform file.