Added loan acceptance logic
Built out the logic for loan acceptance when sent through the offer flow.
This included tests for each of the APR values and for the delenquency
edge case.

Manually ran each of the example cases that are in the seed data to verify
results are as expected.

This completes milestone M1