Built remaining User endpoints
Created the get of a single user and creation of a user endpoints.
This exposed an issue with the id column format of the User object.
It needed to not be a generated column, but had to be a primary column
with the uuid.

Tested manually with the commands in the curl.txt file.