Implemented loan application delete endpoint
Put in the code to soft delete a loan application.  After a bit of looking
it was discovered we could use the softRemove function on TypeORM to
handle the removal.  There was an old bug in TypeORM that didn't allow
for this but has since been remedied.

Again, testing was done manually with commands in curl.txt