Remove direct entity usage from controllers
To provide better separation between the database and the API layers
we need to have layer in between. The common naming is for business
logic to be the Service layer. Following that standard we can push the
use of TypeORMs repositories to be in the the services directory.

We still have the LoanApplication Entity being passed into the
LoanDecider.  The LoanDecider should only be working with data
objects.  This will be a future clean up that will probably mean the
creation of a DAO layer.  Kind of overkill for this application.