Implement missing routes
In V0 we chose to skip two routes because they were primarily
duplicates of other routes. They would use the same code. By moving
some of the code to the service layer it cleaned up the code a
little. It also made it easier to implement the routes.

This is still not optimal as much of what is done is a duplicate
with a couple of calls being different. It is a trade off to not
push the Response down a layer.

A possible solution would be to return an either-like object that
packs the errors together. That should reduce the duplication
though it will open up a new concept for error propigation. Maybe
not a bad thing.